backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
websockets==12.0
openai==1.3.0
pillow==10.1.0
pytesseract==0.3.10
python-dotenv==1.0.0
numpy==1.24.3
pydantic==2.5.0
aiofiles==23.2.1
opencv-python==4.8.1.78